Refactor: Add adjacentField() method and simplify uncover() method
This method returns an adjacent field to a to a field specified by
its coordinates.

The goal is to use the same method to specify adjacent fields in
uncoverAdjacentFields() and surroundMinesWithMarkers() methods or
in any other methods where determining adjacent fields is needed.

uncover() method has been simplified by using adjacentField() within it.